Move grafana vis component to a shared folder and upate its styling.
