+++ src/app/app.component.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
-import { GrafanaVisComponent } from './components/grafana-vis/grafana-vis.component';
+import { GrafanaVisComponent } from './components/shared/grafana-vis/grafana-vis.component';
 import { LoggerTestingModule } from 'ngx-logger/testing';
 import { UIShellModule, TilesModule } from 'carbon-components-angular';
