--- src/app/components/grafana-vis/grafana-vis.component.html
@@ -1,2 +1,0 @@
-<iframe [src]=grafanaUrl width="{{width}}" height="{{height}}" frameborder="0">
-</iframe>
