+++ src/app/components/shared/grafana-vis/grafana-vis.component.html
@@ -1,0 +1,2 @@
+<iframe [src]=grafanaUrl height="{{height}}" class="grafana-vis">
+</iframe>
